% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weather_data.R
\docType{data}
\name{weather_data}
\alias{weather_data}
\title{Melbourne Weather Data}
\format{
A data frame with columns:
\describe{
\item{year}{Year of the observation (integer).}
\item{month}{Month of the observation (integer).}
\item{day}{Day of the observation (integer).}
\item{minimum_temperature_degree_c}{Minimum temperature in degrees Celsius (numeric).}
\item{maximum_temperature_degree_c}{Maximum temperature in degrees Celsius (numeric).}
}
}
\source{
Bureau of Meteorology, Australia
}
\usage{
weather_data
}
\description{
A dataset containing daily minimum and maximum temperatures in Melbourne from 2013 to 2024.
}
\keyword{datasets}
